refactor(EthExchanges): migrate component to Chakra
